Create benford-tester
This algorithm tests the Benford Integrity of " random.randint(x, y) " method in Python's "random" library. Firstly : Benford's law, also called the Newcomb–Benford law, the law of anomalous numbers, or the first-digit law, is an observation about the frequency distribution of leading digits in many real-life sets of numerical data. The law states that in many naturally occurring collections of numbers, the leading digit is likely to be small. For example, in sets that obey the law, the number 1 appears as the leading significant digit about 30% of the time, while 9 appears as the leading significant digit less than 5% of the time. If the digits were distributed uniformly, they would each occur about 11.1% of the time. Benford's law also makes predictions about the distribution of second digits, third digits, digit combinations, and so on.

The hierarchy of digits is :
1- 30.1%	
 
2- 17.6%	
 
3- 12.5%	
 
4- 9.7%	
 
5- 7.9%	
 
6- 6.7%	
 
7- 5.8%	
 
8- 5.1%	
 
9- 4.6%	
So, we can check that law in random.randint(x, y) method in Python. If that method accords Benford's Law, our algorithm should say the repeating digits between 1-1000000 , requires alined like Benford's Law's digit's aline. However I found a disintegrity of Benford's Law in that method.